Don't panicking on LOAD/SAVE, simply return 0. Should fix #20.
On SAVE the procedure doesn't block, it simply does a dummy save on a non-existent tape and return.
Of course, nothing is really saved for now.

On LOAD the procedure waits for data coming from the input port continously reading that port.
This is blocking because the system ROM waits until the load is finished but this never happens.
This is the same behaviour of FUSE when LOADing without starting the tape before. We could intercept
the load event, set a bool loadEvent flag and returning 0 from readPort when that flag is true.
It seems to me an ugly hack though.